Save mask as 8bit png

+++ 3D_loci_simulation/simulate_movie.py
@@ -375,7 +375,7 @@ class SimulateData:
         mask[mask > 0] = 1
         
         with tifffile.TiffWriter(name) as tf:
-            tf.save(np.round(mask).astype(np.uint16))
+            tf.save(np.round(mask).astype(np.uint8))
                 
                 
     def calculate_SNR(self, x0, y0, z0, I):
